requirements.txt
requests
openpyxl
pandas
Office365-REST-Python-Client